backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
langchain~=0.2.16
langchain-openai~=0.1.23
langchain-community~=0.2.16
pydantic>=2.0.0
httpx>=0.27.0
